moved connections and lookup to handler
